README + CRUD methods for Event and Actor services/repositories

+++ src/main/java/com/hackerrank/github/services/EventService.java
@@ -12,5 +12,5 @@ public interface EventService {
 
 	List<Event> getAllEvents();
 
-	List<Event> getAllEvents(int actorID);
+	List<Event> getAllEvents(long actorID);
 }
